hibernate

+++ gui/cashier/forms/EndDayCard.java
@@ -103,10 +103,10 @@ public class EndDayCard
 					JOptionPane.showMessageDialog(endday, "Store not found.");
 					return;
 				}
-				long cashierindex = 0;
+				int cashierindex = 0;
 				try
 				{
-					cashierindex = Long.parseLong(textField_1.getText());
+					cashierindex = Integer.parseInt(textField_1.getText());
 				}
 				catch(NumberFormatException nfe)
 				{
