+++ gui/cashier/forms/StartDayCard.java
@@ -102,10 +102,10 @@ public class StartDayCard
 					JOptionPane.showMessageDialog(startday, "Store not found.");
 					return;
 				}
-				long cashierindex = 0;
+				int cashierindex = 0;
 				try
 				{
-					cashierindex = Long.parseLong(textField_1.getText());
+					cashierindex = Integer.parseInt(textField_1.getText());
 				}
 				catch(NumberFormatException nfe)
 				{
