+++ gui/manager/forms/RemoveCashierCard.java
@@ -104,10 +104,10 @@ public class RemoveCashierCard
 					JOptionPane.showMessageDialog(removecashier, "Store not found.");
 					return;
 				}
-				long cashierindex = 0;
+				int cashierindex = 0;
 				try
 				{
-					cashierindex = Long.parseLong(textField_1.getText());
+					cashierindex = Integer.parseInt(textField_1.getText());
 				}
 				catch(NumberFormatException nfe)
 				{
@@ -127,7 +127,7 @@ public class RemoveCashierCard
 				}
 				try
 				{
-					s.removeCashier((long)cashierindex);
+					s.removeCashier(cashierindex);
 				}
 				catch(IllegalArgumentException iae)
 				{
